feat(hyperlink): 重构超链接组件，提取OG卡片为独立组件
将超链接预览功能中的OG卡片逻辑提取为独立的OGCard组件
简化Hyperlink组件逻辑，移除重复代码
